🖍 Added Transfer UI
- Adding transfer button
- Added new label strings for Transfer
- Polsihed the pretransfer view
- Tech Debt: Added lighter gray to the color palette

+++ loafwallet/CardWalletDetails.swift
@@ -10,13 +10,13 @@ import Foundation
 
 struct CardWalletDetails: Decodable {
     
-    var balance: Float?
+    var balance: Double?
     var ltcAddress: String?
     var createdAt: String?
     var updatedAt: String?
-    var availableBalance: Float?
-    var withdrawableBalance: Float?
-    var spendableBalance: Float?
+    var availableBalance: Double?
+    var withdrawableBalance: Double?
+    var spendableBalance: Double?
     
     enum CodingKeys: String, CodingKey {
         case balance = "balance"
@@ -32,13 +32,13 @@ struct CardWalletDetails: Decodable {
         
         let container = try decoder.container(keyedBy: CodingKeys.self)
         
-        balance = try? container.decode(Float.self, forKey: .balance)
+        balance = try? container.decode(Double.self, forKey: .balance)
         ltcAddress = try? container.decode(String.self, forKey: .ltcAddress)
         createdAt = try? container.decode(String.self, forKey: .createdAt)
         updatedAt = try? container.decode(String.self, forKey: .updatedAt)
-        availableBalance = try? container.decode(Float.self, forKey: .availableBalance)
-        withdrawableBalance = try? container.decode(Float.self, forKey: .withdrawableBalance)
-        spendableBalance = try? container.decode(Float.self, forKey: .spendableBalance)
+        availableBalance = try? container.decode(Double.self, forKey: .availableBalance)
+        withdrawableBalance = try? container.decode(Double.self, forKey: .withdrawableBalance)
+        spendableBalance = try? container.decode(Double.self, forKey: .spendableBalance)
     } 
 }
